fix: resolve webhook publisher type mismatches causing build failures
- Change Webhook.RateLimit config type from int to int64 to match
  watermill's NewThrottle(int64, Duration) signature
- Use message.Publisher interface in webhook publisher struct for
  testability while keeping *kafka.Producer in the FX-wired constructor
- Add NewPublisherWithPublisher constructor for test usage
- Add AsPublisher adapter on InMemoryPubSub to satisfy message.Publisher

+++ internal/config/webhook.go
@@ -13,7 +13,7 @@ type Webhook struct {
 	InitialInterval time.Duration                  `mapstructure:"initial_interval" default:"1s"`
 	MaxInterval     time.Duration                  `mapstructure:"max_interval" default:"10s"`
 	Multiplier      float64                        `mapstructure:"multiplier" default:"2.0"`
-	RateLimit       int                            `mapstructure:"rate_limit" default:"5"`
+	RateLimit       int64                          `mapstructure:"rate_limit" default:"5"`
 	MaxElapsedTime  time.Duration                  `mapstructure:"max_elapsed_time" default:"2m"`
 	Tenants         map[string]TenantWebhookConfig `mapstructure:"tenants"`
 	Svix            Svix                           `mapstructure:"svix_config"`
